requirements.txt
2captcha-python==1.1.3
absl-py==1.2.0
accelerate==0.25.0
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.2.0
aleph-alpha-client==2.14.0
anthropic==0.2.5
anyio==4.2.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bert-score==0.3.13
bitarray==2.7.3
black==22.10.0
blanc==0.2.7
blis==0.7.8
boto3==1.24.89
botocore==1.27.89
bottle==0.12.23
bzip2==1.0.8
ca-certificates==2023.12.12
cachetools==5.2.0
catalogue==2.0.8
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
colorcet==3.0.1
comm==0.1.2
confection==0.1.4
contourpy==1.0.5
crfm-helm==0.2.3
cycler==0.11.0
cymem==2.0.6
cython==0.29.32
dacite==1.6.0
dadapy==0.2.0
datasets==2.5.2
debugpy==1.6.7
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
einops==0.7.0
einsum==0.3.0
emoji==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==0.8.3
fastjsonschema==2.19.1
filelock==3.8.0
flake8==5.0.4
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
gdown==4.4.0
gevent==21.12.0
gin-config==0.5.0
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
huggingface-hub==0.11.0
icetk==0.0.4
identify==2.5.6
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.0
importlib_metadata==7.0.0
iniconfig==1.1.1
ipykernel==6.25.0
ipython==8.15.0
jedi==0.18.1
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
langcodes==3.3.0
ld_impl_linux-64==2.38
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libsodium==1.0.18
libstdcxx-ng==11.2.0
libuuid==1.41.5
llvmlite==0.39.1
lxml==4.9.1
mako==1.2.3
markupsafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
moverscore==1.0.3
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mypy==0.982
mypy-extensions==0.4.3
nbformat==5.9.2
ncurses==6.4
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.7
nodeenv==1.7.0
numba==0.56.4
numpy==1.23.3
openai==0.27.8
openpyxl==3.0.10
openssl==3.0.12
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221003
param==2.0.1
parameterized==0.8.1
parso==0.8.3
pathspec==0.10.1
pathy==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==23.3.1
platformdirs==2.5.2
plotly==5.18.0
pluggy==1.0.0
portalocker==2.5.1
pre-commit==2.20.0
preshed==3.0.7
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.2
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycodestyle==2.9.1
pycparser==2.21
pyct==0.5.0
pydantic==1.8.2
pyemd==0.5.1
pyext==0.7
pyflakes==2.5.0
pygments==2.15.1
pyhocon==0.3.59
pymongo==4.2.0
pyparsing==2.4.7
pysocks==1.7.1
pytest==7.2.0
python==3.10.13
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytrec-eval==0.5
pytz==2022.4
pyyaml==6.0
pyzmq==25.1.0
readline==8.2
referencing==0.32.1
regex==2022.9.13
requests==2.28.1
responses==0.18.0
retrying==1.3.4
rouge-score==0.1.2
rpds-py==0.16.2
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.2.1
sacremoses==0.0.53
safetensors==0.4.1
scaleapi==2.13.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
selenium==4.8.0
sentencepiece==0.1.97
setuptools==68.2.2
simple-slurm==0.2.6
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.3
sqlite==3.41.2
sqlitedict==1.7.0
srsly==2.4.4
stack_data==0.2.0
stanza==1.4.2
summ-eval==0.892
surge-api==1.1.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.1.0
tiktoken==0.3.3
tk==8.6.12
tls-client==0.1.8
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.64.1
traitlets==5.7.1
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.4.2
types-pytz==2022.4.0.0
types-redis==4.3.21.1
types-requests==2.28.11.2
types-tabulate==0.9.0.0
types-urllib3==1.26.25
typing==3.7.4.3
typing-extensions==4.4.0
tzdata==2023d
uncertainty-calibration==0.1.3
undetected-chromedriver==3.2.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
wasabi==0.10.1
wcwidth==0.2.5
websocket-client==1.3.2
websockets==10.4
wheel==0.41.2
wsproto==1.2.0
xlrd==2.0.1
xxhash==3.0.0
xz==5.4.2
yarl==1.8.1
zeromq==4.3.4
zipp==3.11.0
zlib==1.2.13
zope-event==4.5.0
zope-interface==5.4.0
zstandard==0.18.0
